contracts/algorand/pyproject.toml
[tool.poetry]
name = "pact-contracts"
version = "0.1.0"
description = ""
authors = ["krystiansiwek <krystian.siwek@yahoo.com>", "sebastiangula <sebastian.gula@gmail.com>", "lukaszptak <lukasz@ulam.io>", "mateuszwalczak <mateusz.walczak@ulam.io>"]

[tool.poetry.dependencies]
python = "^3.10"
py-algorand-sdk = "^2.0.0"
click = "^8.0.3"
python-dotenv = "^0.19.1"
pytealext = "^3.2.0"
pyteal = "^0.22.0"
beaker-pyteal = "0.5.1"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
pytest-xdist = "^3.1.0"
hypothesis = "^6.65.2"
black = "^23.1.0"
isort = "^5.12.0"
mypy = "0.991"
flake8 = "^6.0.0"
matplotlib = "^3.5.2"
pandas = "^1.4.3"


[tool.black]
exclude = "(.venv|migrations)"
target_version = ["py310"]
line-length = 100

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
skip = [".venv", "migrations"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
